Move query_tc to be TrackingCopy::query
